requirements.txt
Pillow~=9.5.0
streamlit~=1.24.1
numpy~=1.25.1
setuptools~=65.5.1
torch~=2.0.1
boto3~=1.28.2
opencv-python~=4.8.0.74
scikit-image~=0.21.0
python-dotenv~=1.0.0
imageio~=2.31.1
botocore~=1.31.15